+++ src/index.js
@@ -1,5 +1,5 @@
-import Vega from './Vega.js';
-import _createClassFromSpec from './createClassFromSpec.js';
+import Vega from './Vega.jsx';
+import _createClassFromSpec from './createClassFromSpec.jsx';
 
 export default Vega;
 export const createClassFromSpec = _createClassFromSpec;